Renaming module to `_versioning`
